refactor/front-11: rename interval to slider
