+++ mui-atomic/projects/mui-atomic/src/lib/components/index.ts
@@ -8,7 +8,6 @@ export * from './mui-control-base/mui-control-base.component';
 export * from './mui-datepicker/mui-datepicker.component';
 export * from './mui-dialog/index';
 export * from './mui-icon/index';
-export * from './mui-interval/mui-interval';
 export * from './mui-loader/mui-loader';
 export * from './mui-multiselect/mui-multiselect.component';
 export * from './mui-pagination/mui-pagination.component';
@@ -20,6 +19,7 @@ export * from './mui-search/mui-search.component';
 export * from './mui-select-option/mui-select-option.component';
 export * from './mui-select-option/mui-selectable-item';
 export * from './mui-select/mui-select.component';
+export * from './mui-slider/mui-slider';
 export * from './mui-table-body/mui-table-body.component';
 export * from './mui-table-cell/mui-table-cell';
 export * from './mui-table-cell/mui-table-cell.component';
