+++ mui-atomic/projects/mui-atomic/src/lib/components/mui-slider/mui-slider.scss
